Make it better

+++ views/base.jsx
@@ -7,31 +7,33 @@ function Base(props){
   const Page = props.config.page
   function generateFooter(){
     let footer = ''
-    for(let item of props.config.items){
-      footer += `<li>`
-      footer += `<div class='thumb'>`
-      footer += `<a href='/${ item.type }/${ item.id }'>`
-      footer += `<img src='${ item.thumb }'/> `
-      if((item.video)&&(item.video !== '[]')){           
-        footer += `<img class="play-icon" src="/images/play.png"/>`  
-      }   
-      footer += `</a>`
-      footer += `</div>`
-      footer += `<div class="title">`
-      footer += `<a href="/${ props.config.type }/${ item.id }">${ item.title }</a>`         
-      if(props.config.route === '/admin/user'){
-        footer += `<p>${ item.role }</p>`
-      }
-      footer += `<div>${ (new Date(item.postdate)).toLocaleDateString('it-IT') }</div>`
-      footer += `</div>`     
-      footer += `<div class="edit">`       
-      footer += `<a href="${ props.config.route }/edit/${ item.id }"><img src="/images/edit.png"/></a>`       
-      footer += `<a href="${ props.config.route }/delete/${ item.id }"><img src="/images/delete.png"/></a>`   
-      footer += `</div> `   
-      footer += `</li>`                         
-    } 
+    if(props.config.items){
+      for(let item of props.config.items){
+        footer += `<li>`
+        footer += `<div class='thumb'>`
+        footer += `<a href='/${ props.config.type }/${ item.id }'>`
+        footer += `<img src='${ item.thumb }'/> `
+        if((item.video)&&(item.video !== '[]')){           
+          footer += `<img class="play-icon" src="/images/play.png"/>`  
+        }   
+        footer += `</a>`
+        footer += `</div>`
+        footer += `<div class="title">`
+        footer += `<a href="/${ props.config.type }/${ item.id }">${ item.title }</a>`         
+        if(props.config.route === '/admin/user'){
+          footer += `<p>${ item.role }</p>`
+        }
+        footer += `<div>${ (new Date(item.postdate)).toLocaleDateString('it-IT') }</div>`
+        footer += `</div>`     
+        footer += `<div class="edit">`       
+        footer += `<a href="${ props.config.route }/edit/${ item.id }"><img src="/images/edit.png"/></a>`       
+        footer += `<a href="${ props.config.route }/delete/${ item.id }"><img src="/images/delete.png"/></a>`   
+        footer += `</div> `   
+        footer += `</li>`                         
+      } 
 
-    return footer
+      return footer
+    }
   } 
   
   return(
@@ -53,9 +55,10 @@ function Base(props){
 
           <ul class="list" dangerouslySetInnerHTML={{__html: 
             `${ generateFooter() }` }}/>
-          <div class="pagination">
-            <img onClick="paginate(`<%= data.route %>`)" src="/images/load-more.png" />
-          </div>
+          <div class="pagination" dangerouslySetInnerHTML={{__html: `
+            <img onClick="paginate('${ props.config.route }')" src="/images/load-more.png" />
+          `}}/>
+            
           <div class="credit">&copy; 2022 <a href="https://khmerweb.vercel.app/">Khmer Web</a></div>
         </section>
       </body>
